Miscellaneous warning fixes
* src/fontdata.c (compute_text_width): Cast *text indice to int.
* src/rsc_files_hash.h (rsc_file): Replace the modified flag by
a modified_value string.  The initial value will never be modified.
* src/rsc_files.c (set_rsc_file,get_rsc_file): Use modified_value.
* src/rsc_files.gperf: Prototype hash.
* src/timer.c (new_htimer): Fix dmsg format.

+++ src/fontdata.c
@@ -103,6 +103,6 @@ compute_text_width (const fontdata_t *font, const char *text,
   
   for (; *text; ++text)
     if (!ignore_spaces || *text != ' ')
-      width += font->width[*text];
+      width += font->width[(int) *text];
   return width;
 }
